+++ opencascade-hs/src/OpenCascade/HLRBRep/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.HLRBRep.Types
+( Algo
+) where
+
+data Algo